third_party/WebKit/LayoutTests/http/tests/inspector/service-workers/service-worker-manager-expected.txt

Tests the way service worker manager manages targets.

Target added: Main type: 1
Target added: ⚙ service-worker-empty.js type: 4
Target added: ⚙ service-worker-empty.js type: 2
Target removed: ⚙ service-worker-empty.js type: 2
Target removed: ⚙ service-worker-empty.js type: 4
